+++ NativeScript/ffi/objc/napi/ClassMember.mm
@@ -22,7 +22,7 @@
 #include "Util.h"
 #include "Block.h"
 #include "Class.h"
-#include "runtime/NativeScriptException.h"
+#include "runtime/apple/NativeScriptException.h"
 #include "js_native_api.h"
 #include "js_native_api_types.h"
 #include "node_api_util.h"
