+++ NativeScript/runtime/apple/Bundle.h
