+++ NativeScript/runtime/apple/Bundle.mm
