+++ NativeScript/runtime/apple/NativeScript.mm
@@ -1,7 +1,7 @@
 #include "NativeScript.h"
 #include "Runtime.h"
 #include "RuntimeConfig.h"
-#include "runtime/NativeScriptException.h"
+#include "runtime/apple/NativeScriptException.h"
 #include "ffi/objc/shared/Tasks.h"
 #include "js_native_api.h"
 #include "jsr.h"
