+++ NativeScript/runtime/apple/NativeScriptException.mm
@@ -1,4 +1,4 @@
-#include "runtime/NativeScriptException.h"
+#include "runtime/apple/NativeScriptException.h"
 #import <Foundation/Foundation.h>
 #include <sstream>
 #include "js_native_api.h"
@@ -7,7 +7,7 @@
 #include "jsr.h"
 #endif
 #include "native_api_util.h"
-#include "runtime/Runtime.h"
+#include "runtime/apple/Runtime.h"
 
 namespace nativescript {
