+++ NativeScript/runtime/apple/Runtime.h
@@ -6,8 +6,8 @@
 #include <unordered_map>
 
 #include "js_native_api_types.h"
-#include "runtime/SpinLock.h"
-#include "runtime/modules/RuntimeModules.h"
+#include "runtime/apple/SpinLock.h"
+#include "runtime/apple/modules/RuntimeModules.h"
 
 typedef napi_value (*napi_module_init)(napi_env env, napi_value exports);
