+++ NativeScript/runtime/apple/RuntimeConfig.cpp
