+++ NativeScript/runtime/apple/SpinLock.h
