+++ NativeScript/runtime/apple/modules/RuntimeModules.h
@@ -5,11 +5,11 @@
 #include "js_native_api_types.h"
 #include "node/Node.h"
 #include "performance/Performance.h"
-#include "runtime/RuntimeConfig.h"
-#include "runtime/modules/module/ModuleInternal.h"
-#include "runtime/modules/worker/Worker.h"
-#include "url/URL.h"
-#include "url/URLSearchParams.h"
+#include "runtime/apple/RuntimeConfig.h"
+#include "runtime/apple/modules/module/ModuleInternal.h"
+#include "runtime/apple/modules/worker/Worker.h"
+#include "runtime/modules/url/URL.h"
+#include "runtime/modules/url/URLSearchParams.h"
 #include "web/Web.h"
 #ifdef __APPLE__
 #include "app/App.h"
