+++ NativeScript/runtime/apple/modules/app/App.h
