+++ NativeScript/runtime/apple/modules/app/App.mm
