+++ NativeScript/runtime/apple/modules/console/Console.h
