+++ NativeScript/runtime/apple/modules/module/ModuleInternal.cpp
@@ -16,12 +16,12 @@
 #include <unordered_set>
 #include <vector>
 
-#include "runtime/NativeScriptException.h"
+#include "runtime/apple/NativeScriptException.h"
 #include "native_api_util.h"
-#include "runtime/RuntimeConfig.h"
-#include "runtime/Util.h"
-#include "runtime/modules/node/Node.h"
-#include "runtime/modules/web/Web.h"
+#include "runtime/apple/RuntimeConfig.h"
+#include "runtime/apple/Util.h"
+#include "runtime/apple/modules/node/Node.h"
+#include "runtime/apple/modules/web/Web.h"
 
 #ifdef TARGET_ENGINE_V8
 #include "../../napi/v8/v8-module-loader.h"
