+++ NativeScript/runtime/apple/modules/module/ModuleInternal.h
