+++ NativeScript/runtime/apple/modules/node/Node.cpp
