+++ NativeScript/runtime/apple/modules/node/Path.cpp
