+++ NativeScript/runtime/apple/modules/node/VM.cpp
