+++ NativeScript/runtime/apple/modules/performance/Performance.cpp
