+++ NativeScript/runtime/apple/modules/performance/Performance.h
