+++ NativeScript/runtime/apple/modules/worker/Message.h
