+++ NativeScript/runtime/apple/modules/worker/WorkerImpl.mm
@@ -1,16 +1,16 @@
 #include <Foundation/Foundation.h>
 #include <memory>
 #include <string>
-#include "runtime/NativeScriptException.h"
+#include "runtime/apple/NativeScriptException.h"
 #include "js_native_api.h"
 #include "js_native_api_types.h"
 #include "jsr.h"
 #include "native_api_util.h"
 
 #include "WorkerImpl.h"
-#include "runtime/Runtime.h"
-#include "runtime/Util.h"
-#include "runtime/modules/worker/ConcurrentMap.h"
+#include "runtime/apple/Runtime.h"
+#include "runtime/apple/Util.h"
+#include "runtime/apple/modules/worker/ConcurrentMap.h"
 
 namespace nativescript {
